merge main with branch

+++ pages/api/charge/[id].ts
@@ -17,7 +17,7 @@ async function chargeHandler(req: NextApiRequest, res: NextApiResponse) {
     query: { id },
     session
   } = req
-  if (!canUserDo(session, 'READ_CHARGE')) return res.status(403).send(`Can't read this.`)
+  if (!(await canUserDo(session, 'READ_CHARGE'))) return res.status(403).send(`Can't read this.`)
 
   const { success } = idValidation.safeParse(id)
   if (!success) return res.status(404).send(`Id ${id} Not Allowed`)
@@ -39,7 +39,8 @@ async function chargeHandler(req: NextApiRequest, res: NextApiResponse) {
       }
       break
     case 'PUT':
-      if (!canUserDo(session, 'EDIT_CHARGE')) return res.status(403).send(`Can't edit this.`)
+      if (!(await canUserDo(session, 'EDIT_CHARGE')))
+        return res.status(403).send(`Can't edit this.`)
       //actualizamos a UN cargo
       try {
         const charge = await prisma.charge.findFirst({
@@ -60,6 +61,8 @@ async function chargeHandler(req: NextApiRequest, res: NextApiResponse) {
       }
       break
     case 'DELETE':
+      if (!(await canUserDo(session, 'DELETE_CHARGE')))
+        return res.status(403).send(`Can't delete this.`)
       //eliminamos a UN cargo
       try {
         const delCharge = await prisma.charge.delete({ where: { id: Number(id) } })
